extract SceneView's rendered texture cache creation into a hook
